refactor/#93 : Screen 파일 패키지로 이동
- mvi 세팅시 구분 짓기 위함
